Feat: Structure section 4 of the html page
Added necessary structure to this section to
match the example I'm working with. This one I'm
not sure if it will end up working with flexbox.
Usually with two elements that are stacked on top
of one another - i'll have them in a seperate
div compared to another element I want to have
next to those two. This allows me to use
flex-direction in css to have two directions to
control and work with. But time i'm keeping it all
in one flexbox, and I'm going to see if I can just
push a specific item of that flexbox out of the
direction of the other two. This will be a good
test for myself.