Feat: Start work on first section
Changed each section to display as the flexbox
model. This is just initial, but helps with
not having to define it over and over again in the
css. Then I changed section_1 or the first section
to flow in a column direction to match the
example. This section includes both the nav and
the first content section. It may be better to
change this, but at this stage i'm not sure.