Feat: Style the header (navigation)
I've used flexbox on the nav, and will be doing
so for most elements on the page to have them be
responsive in certain ways depending on the
users window size. Having the enclosing
element that holds the various other elements,
like the 'logo' and the <ul> with links allows
me to space the two away from each other, but to
the width of their container. When resizing this
'space-between' will just get smaller.

Note: I could add a margin to either side to stop
the elements touching each other depending
on the window resize.