Feat: Change first section of content to flex
This is the start of sorting out how to get
this content within this section to fit in
terms of its elements and their layout to that
of the example. The image placeholder exists
on the right, and the rest of the content is
to be pushed to the left. But elements such as
the button can't stretch to the full extent of the
flex-container. This needs to be addressed, but
it will take some testing.