Feat: Add some spacing and styling to section 1
Added some spacing between elements. This is
to get it closer to what the example is going
for and help me with adressing certain issues
by being able to see if my use of flexbox isn't
working how I want it to. General styling has been
applied too for visibility of certain elements.