Feat: Change the flex interation in content_1
Made some changes to the use of flexbox. Instead
of using the 'gap' property in css I just used
'justify-content' to get the spacing, this way
it at least scales better than a massive amount
of pixel space used as the gap before the
content gets pushed out of the flexbox.

made a slight change to the html, just shortening
the first paragraph used.