Fix: Fix the html structure for section 2
Section 2 needed to be laid out in a certain way
to allow flexbox to be used for all elements
within it. Because there's a need for having
two adjacent directions the content will be
displayed in I can't do that without seperating
which content I want to control in one way or
the other into their own bigger div elements.