Feat: Change the styling of section 2
I previously changed the html containers to work
with my use case of flexbox for this particular
problem to solve. I had to have those two
divs with content displayed in one direction,
while having the other ones perpendicular to
one another. I was able to get this working in
the end and change some of the gap property
values for spacing.