Fix: Change content 2's section to wrap
Wanted to make this section with these boxes
more responsive in how they act when resizing the
browser window's width. So I had to play around
with the flex-wrap: wrap property and where to
place that with. there were some issues with how
it would space itself, so I had to set it to
'flex: auto' to give it the default grow, shrink,
and flex-basis properties. This has it grow and
shrink equally and base that on the given width
which in this case defaults to its container size.