Feat: Add backround and spacing to content_2
Added a background for seeing how the content
works with wrapping. Also added a gap between the
boxes, this solves the issue with the fact
that space-around is only to the left and right -
when it wraps to a new line on the page there's no
height between the elements (empty space). So gap
solves this.