Fix: Change distance between elements
Distance between elements looked off, so changed
the padding-top to match margins. The way that
I set it up is not effective looking back on this.
I having padding pushing the content down from
the container, and then margin setting where the
next beneath 'starts'. I think there could
be better ways to do this that I don't know yet.